split into otel stuff into seperate module
