+++ flow/otel_metrics/sync_gauges.go
@@ -1,4 +1,4 @@
-package shared
+package otel_metrics
 
 import (
 	"context"
